anisotropicsource: add InvertScreenAndSourceDirection
Add support for a new keyword InvertScreenAndSourceDirection that allows
the user to generate beams emerging from the intensityscreen that are
being forced to go through the anisotropic source, effectively inverting
the default behavior of this device. Ensure that the anisotropicsource
is positioned between intensityscreen and sample/detector for this to
work!

A new example anisotropicsource2 was added to demonstrate this new
feature.